feature/22 refactor: Post의 작성자 위치 필드의 타입을 Address로 변경
