--- src/test/java/com/grasshouse/dorandoran/fixture/AuthorAddressFixture.java
@@ -1,7 +1,0 @@
-package com.grasshouse.dorandoran.fixture;
-
-public class AuthorAddressFixture {
-
-    public static final String AUTHOR_ADDRESS = "부산광역시 연제구 연산동";
-
-}
