+++ src/test/java/com/grasshouse/dorandoran/fixture/PostFixture.java
@@ -1,7 +1,7 @@
 package com.grasshouse.dorandoran.fixture;
 
 import static com.grasshouse.dorandoran.fixture.AddressFixture.ADDRESS;
-import static com.grasshouse.dorandoran.fixture.AuthorAddressFixture.AUTHOR_ADDRESS;
+import static com.grasshouse.dorandoran.fixture.AddressFixture.AUTHOR_ADDRESS;
 import static com.grasshouse.dorandoran.fixture.LocationFixture.JAMSIL_STATION;
 import static com.grasshouse.dorandoran.fixture.MemberFixture.PERSIST_MEMBER;
