+++ src/test/java/com/grasshouse/dorandoran/post/service/PostServiceTest.java
@@ -1,7 +1,7 @@
 package com.grasshouse.dorandoran.post.service;
 
 import static com.grasshouse.dorandoran.fixture.AddressFixture.ADDRESS;
-import static com.grasshouse.dorandoran.fixture.AuthorAddressFixture.AUTHOR_ADDRESS;
+import static com.grasshouse.dorandoran.fixture.AddressFixture.AUTHOR_ADDRESS;
 import static com.grasshouse.dorandoran.fixture.LocationFixture.JAMSIL_STATION;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
